Create hook to make application mutation + sections
